Impl navigate to textfield screen
